* 尝试修正时序问题，修复类型警告

+++ src/jeecs_font_loader.cpp
@@ -91,8 +91,8 @@ jeecs::graphic::character* je_font_get_char(je_font* font, unsigned long chcode)
 
     ch.m_width = x1 - x0;
     ch.m_height = y1 - y0;
-    ch.m_adv_x = real_scalex * advance;
-    ch.m_adv_y = font->m_scale_y;
+    ch.m_adv_x = (int)(real_scalex * (float)advance);
+    ch.m_adv_y = (int)font->m_scale_y;
     ch.m_delta_x = x0;
     ch.m_delta_y = y0;
